Put mathSign into it's own cpp/h file
